Added background block sync, inputtransactions updating service added

+++ BitcoinClient.API/Services/Rpc/ResultEntities/GetTransactionResult.cs
@@ -5,7 +5,7 @@ namespace BitcoinClient.API.Services.Rpc.ResultEntities
     public class GetTransactionResult
     {
         public decimal Fee { get; set; }
-        public int Confirmations { get; set; }
+        public long Confirmations { get; set; }
         public List<GetTransactionResultDetails> Details { get; set; }
         public int Time { get; set; }
     }
